Fix path name generation for Signature artifacts
Currently the filename for artifacts is recreated from their information
like classifier, extension, &c. However, this does not work for
signatures.  Their extension is "asc". The artifact name however is
based on the original artifact. Therefor this information has to be
tracked in order to correctly upload the artifacts' signatures.